+++ src/index.js
@@ -14,6 +14,7 @@ const AssociationsRoutes = require('./routes/associations');
 const StatRoutes = require('./routes/stats');
 const SessionRoute = require('./routes/sessions');
 const ForestRoutes = require('./routes/forest');
+const HealthCheckRoute = require('./routes/healthcheck');
 const Schemas = require('./generators/schemas');
 const SchemaSerializer = require('./serializers/schema');
 const logger = require('./services/logger');
@@ -51,10 +52,17 @@ function getModels() {
 function requireAllModels(modelsDir) {
   if (modelsDir) {
     try {
+      const isJavascriptOrTypescriptFileName = (fileName) =>
+        fileName.endsWith('.js') || (fileName.endsWith('.ts') && !fileName.endsWith('.d.ts'));
+
+      // NOTICE: Ends with `.spec.js`, `.spec.ts`, `.test.js` or `.test.ts`.
+      const isTestFileName = (fileName) => fileName.match(/(?:\.test|\.spec)\.(?:js||ts)$/g);
+
       requireAll({
         dirname: modelsDir,
+        excludeDirs: /^__tests__$/,
         filter: (fileName) =>
-          fileName.endsWith('.js') || (fileName.endsWith('.ts') && !fileName.endsWith('.d.ts')),
+          isJavascriptOrTypescriptFileName(fileName) && !isTestFileName(fileName),
         recursive: true,
       });
     } catch (error) {
@@ -174,6 +182,7 @@ exports.init = (Implementation) => {
     app.use(pathMounted, jwtAuthenticator.unless({ path: pathsPublic }));
   }
 
+  new HealthCheckRoute(app, opts).perform();
   new SessionRoute(app, opts).perform();
 
   // Init
@@ -239,16 +248,21 @@ exports.init = (Implementation) => {
       const collections = _.values(Schemas.schemas);
       configStore.integrator.defineCollections(collections);
 
-      // NOTICE: Check each Smart Action declaration to detect configuration errors.
-      _.each(collections, (collection) => {
-        if (collection.actions) {
-          _.each(collection.actions, (action) => {
-            if (action.fields && !_.isArray(action.fields)) {
+      collections
+        .filter((collection) => collection.actions && collection.actions.length)
+        // NOTICE: Check each Smart Action declaration to detect configuration errors.
+        .forEach((collection) => {
+          const isFieldsInvalid = (action) => action.fields && !Array.isArray(action.fields);
+          collection.actions.forEach((action) => {
+            if (!action.name) {
+              logger.warn(`An unnamed Smart Action of collection "${collection.name}" has been ignored.`);
+            } else if (isFieldsInvalid(action)) {
               logger.error(`Cannot find the fields you defined for the Smart action "${action.name}" of your "${collection.name}" collection. The fields option must be an array.`);
             }
           });
-        }
-      });
+          // NOTICE: Ignore actions without a name.
+          collection.actions = collection.actions.filter((action) => action.name);
+        });
 
       const schemaSerializer = new SchemaSerializer();
       const { options: serializerOptions } = schemaSerializer;
@@ -301,7 +315,7 @@ exports.init = (Implementation) => {
     .then(() => ipWhitelist
       .retrieve(opts.envSecret)
       // NOTICE: An error log (done by the service) is enough in case of retrieval error.
-      .catch(() => {}))
+      .catch(() => { }))
     .catch((error) => {
       logger.error('An error occured while computing the Forest schema. Your application schema cannot be synchronized with Forest. Your admin panel might not reflect your application models definition. ', error);
     });
@@ -343,7 +357,7 @@ exports.collection = (name, opts) => {
     if (opts.searchFields) {
       Schemas.schemas[name].searchFields = opts.searchFields;
     }
-  } else {
+  } else if (opts.fields && opts.fields.length) {
     // NOTICE: Smart Collection definition case
     opts.name = name;
     opts.idField = 'id';
@@ -371,4 +385,4 @@ exports.RecordRemover = require('./services/exposed/record-remover');
 exports.RecordSerializer = require('./services/exposed/record-serializer');
 exports.PermissionMiddlewareCreator = require('./middlewares/permissions');
 
-exports.PUBLIC_ROUTES = ['/', '/sessions', '/sessions-google'];
+exports.PUBLIC_ROUTES = ['/', '/healthcheck', '/sessions', '/sessions-google'];