+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## [Unreleased]
 ### Added
 - Technical - New healthcheck route
+- Linter - Do not lint uncommitted files.
 
 ## RELEASE 5.5.0 - 2020-01-02
 ### Added