+++ CHANGELOG.md
@@ -1,7 +1,7 @@
 # Change Log
 
 ## [Unreleased]
-### Added
+### Fixed
 - Linter - Do not lint uncommitted files.
 
 ## RELEASE 5.5.0 - 2020-01-02