+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 ### Fixed
-- Initialisation - Do not require files that have `__tests__`, `.test.`, or `.spec.` in their filename.
+- Initialization - Filter out test files when requiring models (`__tests__/*`, `*.spec.js`, `*.spec.ts`, `*.test.js` or `*.test.ts`).
 
 ## RELEASE 5.5.0 - 2020-01-02
 ### Added