+++ CHANGELOG.md
@@ -2,7 +2,7 @@
 
 ## [Unreleased]
 ### Added
-- Technical - New healthcheck route
+- Sessions - Distinguish "CORS configuration issue" and "Server down" scenarios in case of liana login error.
 
 ### Fixed
 - Linter - Do not lint uncommitted files.