+++ CHANGELOG.md
@@ -1,6 +1,10 @@
 # Change Log
 
 ## [Unreleased]
+### Fixed
+- Schema - Schemas having fields with escaped characters are now properly saved as valid JSON
+- Tests - Add tests for json prettyPrint
+- Tests - Add tests for schema file updater
 
 ## RELEASE 5.6.0 - 2020-01-14
 ### Added